rename fixtures.pageserver.types
