+++ test_runner/fixtures/pageserver/remote_storage.py
@@ -8,7 +8,7 @@ from typing import Any, List, Tuple
 
 from fixtures.common_types import TenantId, TimelineId
 from fixtures.neon_fixtures import NeonEnv, Pagectl
-from fixtures.pageserver.types import (
+from fixtures.pageserver.common_types import (
     InvalidFileName,
     parse_layer_file_name,
 )
